Connect With Python API using axios in BlogCard
