+++ src/App.jsx
@@ -5,7 +5,7 @@ import Home from "./components/Home";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
 import Project from "./components/Project";
-import BlogPage from "./components/Blog";
+import BlogIndex from "./components/Blog/BlogIndex";
 import AboutUs from "./components/AboutUs";
 // import Hooks from "./components/tryingHooks"
 
@@ -15,15 +15,13 @@ export default class App extends Component {
       <Router>
         <div>
           <Navbar />
-          <Route exact path="/home" component={Home} />
+          <Route exact path="/" component={Home} />
           <Route path="/projects" component={Project} />
-          <Route path="/study-guides" component={BlogPage} />
+          <Route path="/study-guides" component={BlogIndex} />
           <Route path="/about-us" component={AboutUs} />
           <Footer />
         </div>
       </Router>
-      
-     
     );
   }
 }
